Cleanup long long/BigInt inconsistency
git-svn-id: svn+ssh://keith.cs.st-and.ac.uk/mnt/raid/svn/repositories/minion-private/public/trunk@1788 477b24dd-e91e-0410-aa49-980f3df0729b

+++ minion/preprocess.cpp
@@ -22,7 +22,7 @@ void PropogateCSP(StateObj* stateObj, PropagationLevel preprocessLevel, vector<A
   if(preprocessLevel == PropLevel_GAC)
     return;
   
-  long long lits = lit_count(vars);
+  BigInt lits = lit_count(vars);
   bool bounds_check = (preprocessLevel == PropLevel_SACBounds) ||
   (preprocessLevel == PropLevel_SSACBounds);
   clock_t start_SAC_time = clock();
